Fixing compile errors for nvhpc

+++ src/mesh/MeshUtilities.cpp
@@ -313,9 +313,9 @@ inline double cellVolume3D( const AMP::Geometry::Geometry &geom,
                     double z      = x0[2] + 0.5 * k * dx[2];
                     double xyz[3] = { x, y, z };
                     bool in3[8]   = { in2[i][j][k],         in2[i + 1][j][k],
-                                    in2[i][j + 1][k],     in2[i + 1][j + 1][k],
-                                    in2[i][j][k + 1],     in2[i + 1][j][k + 1],
-                                    in2[i][j + 1][k + 1], in2[i + 1][j + 1][k + 1] };
+                                      in2[i][j + 1][k],     in2[i + 1][j + 1][k],
+                                      in2[i][j][k + 1],     in2[i + 1][j][k + 1],
+                                      in2[i][j + 1][k + 1], in2[i + 1][j + 1][k + 1] };
                     volume += cellVolume3D( geom, xyz, dx2, in3, tol );
                 }
             }
@@ -378,8 +378,8 @@ Array<double> volumeOverlap( const AMP::Geometry::Geometry &geom, const std::vec
         // Get the bounding box
         auto [lb, ub] = geom.box();
         double dx[3]  = { ( ub[0] - lb[0] ) / N[0],
-                         ( ub[1] - lb[1] ) / N[1],
-                         ( ub[2] - lb[2] ) / N[2] };
+                          ( ub[1] - lb[1] ) / N[1],
+                          ( ub[2] - lb[2] ) / N[2] };
         // Get the volume for each cell
         volume.resize( N[0], N[1], N[2] );
         volume.fill( 0 );
@@ -404,9 +404,9 @@ Array<double> volumeOverlap( const AMP::Geometry::Geometry &geom, const std::vec
                     double x          = lb[0] + i * dx[0];
                     double xyz[3]     = { x, y, z };
                     bool in[8]        = { inside( i, j, k ),         inside( i + 1, j, k ),
-                                   inside( i, j + 1, k ),     inside( i + 1, j + 1, k ),
-                                   inside( i, j, k + 1 ),     inside( i + 1, j, k + 1 ),
-                                   inside( i, j + 1, k + 1 ), inside( i + 1, j + 1, k + 1 ) };
+                                          inside( i, j + 1, k ),     inside( i + 1, j + 1, k ),
+                                          inside( i, j, k + 1 ),     inside( i + 1, j, k + 1 ),
+                                          inside( i, j + 1, k + 1 ), inside( i + 1, j + 1, k + 1 ) };
                     volume( i, j, k ) = cellVolume3D( geom, xyz, dx, in, tol );
                 }
             }
@@ -423,7 +423,7 @@ Array<double> volumeOverlap( const AMP::Geometry::Geometry &geom, const std::vec
  ********************************************************/
 ElementFinder::ElementFinder( std::shared_ptr<AMP::Mesh::Mesh> mesh )
     : d_mesh( mesh ),
-      d_pos_hash( -1 ),
+      d_pos_hash( static_cast<size_t>( -1 ) ),
       d_elements( mesh->getIterator( mesh->getGeomType() ) ),
       d_dist( std::numeric_limits<double>::quiet_NaN() )
 {
@@ -431,7 +431,7 @@ ElementFinder::ElementFinder( std::shared_ptr<AMP::Mesh::Mesh> mesh )
 }
 ElementFinder::ElementFinder( std::shared_ptr<AMP::Mesh::Mesh> mesh, AMP::Mesh::MeshIterator it )
     : d_mesh( mesh ),
-      d_pos_hash( -1 ),
+      d_pos_hash( static_cast<size_t>( -1 ) ),
       d_elements( std::move( it ) ),
       d_dist( std::numeric_limits<double>::quiet_NaN() )
 {
